requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
python-multipart
scikit-learn
python-dotenv
rasa
requests